requirements.txt
# SQL Analyst Agent - Requirements

# Core dependencies
langgraph>=0.0.1
langchain>=0.1.0
langchain-groq>=0.0.1
groq>=0.0.1

# Database
mysql-connector-python>=9.0.0

# Web Server
fastapi>=0.100.0
uvicorn>=0.20.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
pydantic>=2.0.0

# Optional: Web Interface
gradio>=4.0.0

# Development
pytest>=7.0.0
